Frontend. Refactor place creation logic into PlaceForm component which allows both creating and editing places
