+++ fotolyrik_frontend/pages/places/[id]/index.vue
